feat(github): PR reviews, checks, inline comments & file-content APIs
Add Tauri commands for PR reviews, review comments (create/reply),
combined CI checks, and file content at a commit SHA (Contents API);
add ETag/If-None-Match conditional caching to the REST client so warm
list/detail refreshes return 304 without counting against rate limits.

Split the monolithic github commands module into per-domain files
(repos, pulls, issues, credentials, shared) behind unchanged public
command paths.

Pre-commit hook ran. Total eslint: 0, total circular: 0
