+++ src-tauri/src/commands/handler_list.inc
@@ -170,12 +170,19 @@ integrations::github::commands::github_list_repos,
 integrations::github::commands::github_search_repos,
 integrations::github::commands::github_list_branches,
 integrations::github::commands::github_create_branch,
+integrations::github::commands::github_get_content,
 integrations::github::commands::github_create_pr,
 integrations::github::commands::github_find_pull_request,
 integrations::github::commands::github_list_open_prs,
 integrations::github::commands::github_get_pr,
 integrations::github::commands::github_list_pr_commits,
 integrations::github::commands::github_list_pr_files,
+integrations::github::commands::github_list_pr_reviews,
+integrations::github::commands::github_list_pr_review_comments,
+integrations::github::commands::github_create_pr_review,
+integrations::github::commands::github_create_pr_review_comment,
+integrations::github::commands::github_reply_pr_review_comment,
+integrations::github::commands::github_get_checks,
 integrations::github::commands::github_clone_repo,
 integrations::github::commands::github_git_credential_for_remote,
 integrations::github::commands::github_check_token,
